Выносит масштабирование в отдельный модуль, добавляет сброс масштаба при отправке формы
